Extract Spotify media browsing into a module
