+++ homeassistant/components/spotify/const.py
@@ -1,4 +1,11 @@
 """Define constants for the Spotify integration."""
+from homeassistant.components.media_player.const import (
+    MEDIA_TYPE_ALBUM,
+    MEDIA_TYPE_ARTIST,
+    MEDIA_TYPE_EPISODE,
+    MEDIA_TYPE_PLAYLIST,
+    MEDIA_TYPE_TRACK,
+)
 
 DOMAIN = "spotify"
 
@@ -24,3 +31,13 @@ SPOTIFY_SCOPES = [
 ]
 
 MEDIA_PLAYER_PREFIX = "spotify://"
+MEDIA_TYPE_SHOW = "show"
+
+PLAYABLE_MEDIA_TYPES = [
+    MEDIA_TYPE_PLAYLIST,
+    MEDIA_TYPE_ALBUM,
+    MEDIA_TYPE_ARTIST,
+    MEDIA_TYPE_EPISODE,
+    MEDIA_TYPE_SHOW,
+    MEDIA_TYPE_TRACK,
+]
